routine_ai/server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.8.2
